new_test_app_react/Backend/pyproject.toml
[tool.poetry]
name = "business-insights-agent"
version = "0.1.0"
description = "CopilotKit agent for business intelligence"
authors = ["Your Team <team@company.com>"]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
copilotkit = "0.1.48"
langchain = ">=0.1.0"
langchain-core = ">=0.1.5"
langchain-openai = ">=0.0.1"
langgraph = "^0.3.25"
fastapi = "0.115.12"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
agent = "agent.agent:main"
